Refactor MoPub Adapters
