+++ hybid.adapters.mopub/src/main/java/net/pubnative/lite/adapters/mopub/PNLiteMoPubBannerCustomEvent.java
@@ -22,99 +22,8 @@
 //
 package net.pubnative.lite.adapters.mopub;
 
-import android.content.Context;
-import android.view.View;
-
-import com.mopub.mobileads.CustomEventBanner;
-import com.mopub.mobileads.MoPubErrorCode;
-
-import net.pubnative.lite.sdk.PNLite;
-import net.pubnative.lite.sdk.banner.presenter.BannerPresenter;
-import net.pubnative.lite.sdk.banner.presenter.BannerPresenterFactory;
-import net.pubnative.lite.sdk.models.Ad;
-import net.pubnative.lite.sdk.utils.Logger;
-
-import java.util.Map;
-
-/**
- * Created by erosgarciaponte on 12.01.18.
- */
-
-public class PNLiteMoPubBannerCustomEvent extends CustomEventBanner implements BannerPresenter.Listener {
-    private static final String TAG = PNLiteMoPubBannerCustomEvent.class.getSimpleName();
-
-    private static final String ZONE_ID_KEY = "pn_zone_id";
-    private CustomEventBannerListener mBannerListener;
-    private BannerPresenter mBannerPresenter;
-
-    @Override
-    protected void loadBanner(Context context,
-                              CustomEventBannerListener customEventBannerListener,
-                              Map<String, Object> localExtras,
-                              Map<String, String> serverExtras) {
-
-        if (customEventBannerListener == null) {
-            Logger.e(TAG, "customEventBannerListener is null");
-            return;
-        }
-        mBannerListener = customEventBannerListener;
-
-        String zoneIdKey;
-        if (localExtras.containsKey(ZONE_ID_KEY)) {
-            zoneIdKey = (String) localExtras.get(ZONE_ID_KEY);
-        } else if (serverExtras.containsKey(ZONE_ID_KEY)) {
-            zoneIdKey = serverExtras.get(ZONE_ID_KEY);
-        } else {
-            Logger.e(TAG, "Could not find zone id value in CustomEventBanner localExtras or serverExtras");
-            mBannerListener.onBannerFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
-            return;
-        }
-
-        final Ad ad = PNLite.getAdCache().remove(zoneIdKey);
-        if (ad == null) {
-            Logger.e(TAG, "Could not find an ad in the cache for zone id with key: " + zoneIdKey);
-            mBannerListener.onBannerFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
-            return;
-        }
-
-        mBannerPresenter = new BannerPresenterFactory(context).createBannerPresenter(ad, this);
-        if (mBannerPresenter == null) {
-            Logger.e(TAG, "Could not create valid banner presenter");
-            mBannerListener.onBannerFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
-            return;
-        }
-
-        mBannerPresenter.load();
-    }
-
-    @Override
-    protected void onInvalidate() {
-        if (mBannerPresenter != null) {
-            mBannerPresenter.stopTracking();
-            mBannerPresenter.destroy();
-            mBannerPresenter = null;
-        }
-    }
-
-    @Override
-    public void onBannerLoaded(BannerPresenter bannerPresenter, View banner) {
-        if (mBannerListener != null) {
-            mBannerListener.onBannerLoaded(banner);
-            mBannerPresenter.startTracking();
-        }
-    }
-
-    @Override
-    public void onBannerClicked(BannerPresenter bannerPresenter) {
-        if (mBannerListener != null) {
-            mBannerListener.onBannerClicked();
-        }
-    }
-
-    @Override
-    public void onBannerError(BannerPresenter bannerPresenter) {
-        if (mBannerListener != null) {
-            mBannerListener.onBannerFailed(MoPubErrorCode.INTERNAL_ERROR);
-        }
-    }
+public class PNLiteMoPubBannerCustomEvent extends HyBidMoPubBannerCustomEvent {
+    /*
+     *  This class is kept for backwards compatibility.
+     */
 }
