+++ hybid.adapters.mopub/src/main/java/net/pubnative/lite/adapters/mopub/mediation/HyBidMediationMRectCustomEvent.java
@@ -1,0 +1,112 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2018 PubNative GmbH
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+package net.pubnative.lite.adapters.mopub.mediation;
+
+import android.content.Context;
+
+import com.mopub.mobileads.CustomEventBanner;
+import com.mopub.mobileads.MoPubErrorCode;
+
+import net.pubnative.lite.sdk.HyBid;
+import net.pubnative.lite.sdk.utils.Logger;
+import net.pubnative.lite.sdk.views.PNAdView;
+import net.pubnative.lite.sdk.views.PNMRectAdView;
+
+import java.util.Map;
+
+public class HyBidMediationMRectCustomEvent extends CustomEventBanner implements PNAdView.Listener {
+    private static final String TAG = HyBidMediationMRectCustomEvent.class.getSimpleName();
+
+    private static final String APP_TOKEN_KEY = "pn_app_token";
+    private static final String ZONE_ID_KEY = "pn_zone_id";
+    private CustomEventBannerListener mBannerListener;
+    private PNMRectAdView mMRectView;
+
+    @Override
+    protected void loadBanner(Context context,
+                              CustomEventBannerListener customEventBannerListener,
+                              Map<String, Object> localExtras,
+                              Map<String, String> serverExtras) {
+        if (customEventBannerListener == null) {
+            Logger.e(TAG, "customEventBannerListener is null");
+            return;
+        }
+
+        mBannerListener = customEventBannerListener;
+
+        String zoneId;
+        String appToken;
+        if (serverExtras.containsKey(ZONE_ID_KEY) && serverExtras.containsKey(APP_TOKEN_KEY)) {
+            zoneId = serverExtras.get(ZONE_ID_KEY);
+            appToken = serverExtras.get(APP_TOKEN_KEY);
+        } else {
+            Logger.e(TAG, "Could not find the required params in CustomEventBanner serverExtras");
+            mBannerListener.onBannerFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+            return;
+        }
+
+        if (appToken == null || !appToken.equals(HyBid.getAppToken())) {
+            Logger.e(TAG, "The provided app token doesn't match the one used to initialise HyBid");
+            mBannerListener.onBannerFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+            return;
+        }
+
+        mMRectView = new PNMRectAdView(context);
+        mMRectView.load(zoneId, this);
+    }
+
+    @Override
+    protected void onInvalidate() {
+        if (mMRectView != null) {
+            mMRectView.destroy();
+            mMRectView = null;
+        }
+    }
+
+    //------------------------------------ PNAdView Callbacks --------------------------------------
+    @Override
+    public void onAdLoaded() {
+        if (mBannerListener != null) {
+            mBannerListener.onBannerLoaded(mMRectView);
+        }
+    }
+
+    @Override
+    public void onAdLoadFailed(Throwable error) {
+        if (mBannerListener != null) {
+            mBannerListener.onBannerFailed(MoPubErrorCode.NETWORK_NO_FILL);
+        }
+    }
+
+    @Override
+    public void onAdImpression() {
+
+    }
+
+    @Override
+    public void onAdClick() {
+        if (mBannerListener != null) {
+            mBannerListener.onBannerClicked();
+        }
+    }
+}
